+++ sdk/metricsadvisor/azure-ai-metricsadvisor/tests/base_testcase.py
@@ -8,7 +8,8 @@
 import datetime
 from devtools_testutils import AzureTestCase
 from azure_devtools.scenario_tests import (
-    ReplayableTest
+    ReplayableTest,
+    create_random_name
 )
 
 from azure.ai.metricsadvisor import (
@@ -174,7 +175,7 @@ class TestMetricsAdvisorAdministrationClientBase(AzureTestCase):
                                                                MetricsAdvisorKeyCredential(subscription_key, api_key))
 
     def _create_data_feed(self, name):
-        name = self.create_random_name(name)
+        name = create_random_name(name)
         return self.admin_client.create_data_feed(
             DataFeed(
                 name=name,
@@ -204,7 +205,7 @@ class TestMetricsAdvisorAdministrationClientBase(AzureTestCase):
 
     def _create_data_feed_and_anomaly_detection_config(self, name):
         data_feed = self._create_data_feed(name)
-        detection_config_name = self.create_random_name(name)
+        detection_config_name = create_random_name(name)
         detection_config = self.admin_client.create_metric_anomaly_detection_configuration(
             AnomalyDetectionConfiguration(
                 name=detection_config_name,
@@ -225,7 +226,7 @@ class TestMetricsAdvisorAdministrationClientBase(AzureTestCase):
         return detection_config, data_feed
 
     def _create_data_feed_for_update(self, name):
-        data_feed_name = self.create_random_name(name)
+        data_feed_name = create_random_name(name)
         return self.admin_client.create_data_feed(
             DataFeed(
                 name=data_feed_name,
@@ -273,7 +274,7 @@ class TestMetricsAdvisorAdministrationClientBase(AzureTestCase):
 
     def _create_anomaly_alert_config_for_update(self, name):
         detection_config, data_feed = self._create_data_feed_and_anomaly_detection_config(name)
-        alert_config_name = self.create_random_name(name)
+        alert_config_name = create_random_name(name)
         alert_config = self.admin_client.create_anomaly_alert_configuration(
             AnomalyAlertConfiguration(
                 name=alert_config_name,
@@ -331,7 +332,7 @@ class TestMetricsAdvisorAdministrationClientBase(AzureTestCase):
 
     def _create_detection_config_for_update(self, name):
         data_feed = self._create_data_feed(name)
-        detection_config_name = self.create_random_name("testupdated")
+        detection_config_name = create_random_name("testupdated")
         detection_config = self.admin_client.create_metric_anomaly_detection_configuration(
             AnomalyDetectionConfiguration(
                 name=detection_config_name,