+++ sdk/servicebus/azure-servicebus/tests/async_tests/test_sessions_async.py
@@ -16,7 +16,7 @@ from datetime import datetime, timedelta
 from uamqp.errors import VendorLinkDetach
 from azure.servicebus.aio import ServiceBusClient, ReceivedMessage, AutoLockRenewer
 from azure.servicebus._common.message import Message, PeekedMessage
-from azure.servicebus._common.constants import ReceiveMode, NEXT_AVAILABLE, SubQueue
+from azure.servicebus._common.constants import ReceiveMode, NEXT_AVAILABLE_SESSION, SubQueue
 from azure.servicebus._common.utils import utc_now
 from azure.servicebus.exceptions import (
     ServiceBusConnectionError,
@@ -62,7 +62,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
             with pytest.raises(ServiceBusConnectionError):
                 await sb_client.get_queue_receiver(servicebus_queue.name, max_wait_time=5)._open_with_retry()
 
-            session = sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5)
+            session = sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5)
             count = 0
             async for message in session:
                 print_message(_logger, message)
@@ -91,7 +91,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                     await sender.send_messages(message)
 
             messages = []
-            session = sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, receive_mode=ReceiveMode.ReceiveAndDelete, max_wait_time=5)
+            session = sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, receive_mode=ReceiveMode.ReceiveAndDelete, max_wait_time=5)
             async for message in session:
                 messages.append(message)
                 assert session_id == session.session.session_id
@@ -107,7 +107,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
             time.sleep(30)
 
             messages = []
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, receive_mode=ReceiveMode.ReceiveAndDelete, max_wait_time=5) as session:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, receive_mode=ReceiveMode.ReceiveAndDelete, max_wait_time=5) as session:
                 async for message in session:
                     messages.append(message)
             assert len(messages) == 0
@@ -129,7 +129,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                     await sender.send_messages(message)
 
             messages = []
-            session = sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5)
+            session = sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5)
             async with session:
                 async for message in session:
                     assert session_id == session.session.session_id
@@ -164,7 +164,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
         async with ServiceBusClient.from_connection_string(
             servicebus_namespace_connection_string, logging_enable=False) as sb_client:
 
-            session = sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE, max_wait_time=5)
+            session = sb_client.get_queue_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE_SESSION, max_wait_time=5)
             with pytest.raises(NoActiveSession):
                 await session._open_with_retry()
 
@@ -180,7 +180,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
 
             session_id = str(uuid.uuid4())
             messages = []
-            session = sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, receive_mode=ReceiveMode.ReceiveAndDelete, max_wait_time=5)
+            session = sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, receive_mode=ReceiveMode.ReceiveAndDelete, max_wait_time=5)
             async with session:
                 async for message in session:
                     messages.append(message)
@@ -205,7 +205,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                     await sender.send_messages(message)
 
             count = 0
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
                 async for message in session:
                     deferred_messages.append(message.sequence_number)
                     print_message(_logger, message)
@@ -214,7 +214,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
 
             assert count == 10
 
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
                 deferred = await session.receive_deferred_messages(deferred_messages)
                 assert len(deferred) == 10
                 for message in deferred:
@@ -242,7 +242,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                     await sender.send_messages(message)
 
             count = 0
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
                 async for message in session:
                     deferred_messages.append(message.sequence_number)
                     print_message(_logger, message)
@@ -251,7 +251,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
 
             assert count == 10
 
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
                 deferred = await session.receive_deferred_messages(deferred_messages)
                 assert len(deferred) == 10
                 for message in deferred:
@@ -288,7 +288,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                     await sender.send_messages(message)
 
             count = 0
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
                 async for message in session:
                     deferred_messages.append(message.sequence_number)
                     print_message(_logger, message)
@@ -296,7 +296,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                     await message.defer()
 
             assert count == 10
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ReceiveMode.ReceiveAndDelete) as session:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ReceiveMode.ReceiveAndDelete) as session:
                 deferred = await session.receive_deferred_messages(deferred_messages)
                 assert len(deferred) == 10
                 for message in deferred:
@@ -323,7 +323,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                     message = Message("Deferred message no. {}".format(i), session_id=session_id)
                     await sender.send_messages(message)
 
-            session = sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5)
+            session = sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5)
             count = 0
             async for message in session:
                 deferred_messages.append(message.sequence_number)
@@ -347,7 +347,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
             servicebus_namespace_connection_string, logging_enable=False) as sb_client:
 
             session_id = str(uuid.uuid4())
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, prefetch_count=10) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, prefetch_count=10) as receiver:
 
                 async with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     for i in range(10):
@@ -399,7 +399,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                     message = Message("Test message no. {}".format(i), session_id=session_id_2)
                     await sender.send_messages(message)
 
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                 messages = await receiver.peek_messages(5)
                 assert len(messages) == 5
                 assert all(isinstance(m, PeekedMessage) for m in messages)
@@ -408,7 +408,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                     with pytest.raises(AttributeError):
                         message.complete()
 
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id_2) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id_2) as receiver:
                 messages = await receiver.peek_messages(5)
                 assert len(messages) == 3
 
@@ -422,7 +422,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
             servicebus_namespace_connection_string, logging_enable=False) as sb_client:
 
             session_id = str(uuid.uuid4())
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, max_wait_time=5, session_id=session_id) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, max_wait_time=5, session_id=session_id) as receiver:
                 async with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     for i in range(5):
                         message = Message("Test message no. {}".format(i), session_id=session_id)
@@ -449,7 +449,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
             session_id = str(uuid.uuid4())
             messages = []
             locks = 3
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, prefetch_count=10) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, prefetch_count=10) as receiver:
                 async with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     for i in range(locks):
                         message = Message("Test message no. {}".format(i), session_id=session_id)
@@ -500,7 +500,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
 
             renewer = AutoLockRenewer()
             messages = []
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ReceiveMode.PeekLock, prefetch_count=20) as session:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ReceiveMode.PeekLock, prefetch_count=20) as session:
                 renewer.register(session.session, timeout=60)
                 print("Registered lock renew thread", session.session.locked_until_utc, utc_now())
                 with pytest.raises(SessionLockExpired):
@@ -535,7 +535,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
             renewer._renew_period = 1
             session = None
 
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ReceiveMode.PeekLock, prefetch_count=10) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ReceiveMode.PeekLock, prefetch_count=10) as receiver:
                 session = receiver.session
                 renewer.register(session, timeout=5, on_lock_renew_failure=lock_lost_callback)
             await asyncio.sleep(max(0,(session.locked_until_utc - utc_now()).total_seconds()+1)) # If this pattern repeats make sleep_until_expired_async
@@ -561,7 +561,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                 message.session_id = session_id
                 await sender.send_messages(message)
 
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                 messages = await receiver.receive_messages(max_wait_time=10)
                 assert len(messages) == 1
 
@@ -585,7 +585,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                 message.session_id = session_id
                 await sender.send_messages(message)
 
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                 messages = await receiver.receive_messages(max_wait_time=10)
                 assert len(messages) == 1
                 print_message(_logger, messages[0])
@@ -600,7 +600,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                 with pytest.raises(SessionLockExpired):
                     await receiver.session.renew_lock()
 
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                 messages = await receiver.receive_messages(max_wait_time=30)
                 assert len(messages) == 1
                 print_message(_logger, messages[0])
@@ -629,7 +629,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
 
             messages = []
             renewer = AutoLockRenewer()
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                 renewer.register(receiver.session, timeout=140)
                 messages.extend(await receiver.receive_messages(max_wait_time=120))
                 messages.extend(await receiver.receive_messages(max_wait_time=5))
@@ -669,7 +669,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                 assert len(tokens) == 2
 
             renewer = AutoLockRenewer()
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, prefetch_count=20) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, prefetch_count=20) as receiver:
                 renewer.register(receiver.session, timeout=140)
                 messages.extend(await receiver.receive_messages(max_wait_time=120))
                 messages.extend(await receiver.receive_messages(max_wait_time=5))
@@ -705,7 +705,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
 
             renewer = AutoLockRenewer()
             messages = []
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                 renewer.register(receiver.session, timeout=140)
                 messages.extend(await receiver.receive_messages(max_wait_time=120))
                 messages.extend(await receiver.receive_messages(max_wait_time=5))
@@ -734,7 +734,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                     message = Message("Handler message no. {}".format(i), session_id=session_id)
                     await sender.send_messages(message)
 
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
                 assert await session.session.get_state(timeout=5) == None
                 await session.session.set_state("first_state", timeout=5)
                 count = 0
@@ -766,10 +766,10 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                         message = Message("Test message no. {}".format(i), session_id=session)
                         await sender.send_messages(message)
             for session in sessions:
-                async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session) as receiver:
+                async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session) as receiver:
                     await receiver.session.set_state("SESSION {}".format(session))
 
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE, max_wait_time=5, receive_mode=ReceiveMode.PeekLock) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE_SESSION, max_wait_time=5, receive_mode=ReceiveMode.PeekLock) as receiver:
                 current_sessions = await receiver.list_sessions(updated_since=start_time)
                 assert len(current_sessions) == 5
                 assert current_sessions == sessions
@@ -796,7 +796,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                         message = Message("Test message no. {}".format(i), session_id=session)
                         await sender.send_messages(message)
             for session in sessions:
-                async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session) as receiver:
+                async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session) as receiver:
                     await receiver.session.set_state("SESSION {}".format(session))
 
             current_sessions = await sb_client.list_sessions(updated_since=start_time)
@@ -816,7 +816,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
         async def message_processing(sb_client):
             while True:
                 try:
-                    async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE, max_wait_time=5) as session:
+                    async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE_SESSION, max_wait_time=5) as session:
                         async for message in session:
                             print("Message: {}".format(message))
                             messages.append(message)
@@ -857,7 +857,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
                 message = Message(b"Sample topic message", session_id='test_session')
                 await sender.send_messages(message)
 
-            async with sb_client.get_subscription_session_receiver(
+            async with sb_client.get_subscription_receiver(
                 topic_name=servicebus_topic.name,
                 subscription_name=servicebus_subscription.name,
                 session_id='test_session',
@@ -893,7 +893,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
             assert not receiver._handler
 
             # Then let's try a case we can recover from to make sure everything works on reestablishment.
-            receiver = sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE)
+            receiver = sb_client.get_queue_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE_SESSION)
             with pytest.raises(NoActiveSession):
                 await receiver._open_with_retry()
 
@@ -901,7 +901,7 @@ class ServiceBusAsyncSessionTests(AzureMgmtTestCase):
             async with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                 await sender.send_messages(Message("test session sender", session_id=session_id))
 
-            async with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE, max_wait_time=5) as receiver:
+            async with sb_client.get_queue_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE_SESSION, max_wait_time=5) as receiver:
                 messages = []
                 async for message in receiver:
                     messages.append(message)