+++ sdk/servicebus/azure-servicebus/tests/test_sessions.py
@@ -15,7 +15,7 @@ from datetime import datetime, timedelta
 
 from azure.servicebus import ServiceBusClient, AutoLockRenewer
 from azure.servicebus._common.message import Message, PeekedMessage, ReceivedMessage
-from azure.servicebus._common.constants import ReceiveMode, NEXT_AVAILABLE, SubQueue
+from azure.servicebus._common.constants import ReceiveMode, NEXT_AVAILABLE_SESSION, SubQueue
 from azure.servicebus._common.utils import utc_now
 from azure.servicebus.exceptions import (
     ServiceBusConnectionError,
@@ -53,7 +53,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
 
             session_id = str(uuid.uuid4())
             sender = sb_client.get_queue_sender(servicebus_queue.name)
-            session = sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5)
+            session = sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5)
 
             with sender, session:
                 for i in range(3):
@@ -118,7 +118,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                     sender.send_messages(message)
 
             messages = []
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, 
+            with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               session_id=session_id, 
                                               receive_mode=ReceiveMode.ReceiveAndDelete, 
                                               max_wait_time=5) as session:
@@ -134,7 +134,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             time.sleep(30)
 
             messages = []
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, receive_mode=ReceiveMode.ReceiveAndDelete, max_wait_time=5) as session:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, receive_mode=ReceiveMode.ReceiveAndDelete, max_wait_time=5) as session:
                 for message in session:
                     messages.append(message)
                 assert len(messages) == 0
@@ -155,7 +155,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                     sender.send_messages(message)
 
             messages = []
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
                 for message in session:
                     assert session_id == session.session.session_id 
                     assert session_id == message.session_id
@@ -188,8 +188,8 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
         with ServiceBusClient.from_connection_string(
             servicebus_namespace_connection_string, logging_enable=False) as sb_client:
             with pytest.raises(NoActiveSession):
-                with sb_client.get_queue_session_receiver(servicebus_queue.name, 
-                                                  session_id=NEXT_AVAILABLE, 
+                with sb_client.get_queue_receiver(servicebus_queue.name, 
+                                                  session_id=NEXT_AVAILABLE_SESSION, 
                                                   max_wait_time=5) as session:
                         session.open()
 
@@ -218,7 +218,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             assert not receiver._handler
 
             # Then let's try a case we can recover from to make sure everything works on reestablishment.
-            receiver = sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE)
+            receiver = sb_client.get_queue_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE_SESSION)
             with pytest.raises(NoActiveSession):
                 receiver._open_with_retry()
 
@@ -226,7 +226,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                 sender.send_messages(Message("test session sender", session_id=session_id))
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE, max_wait_time=5) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE_SESSION, max_wait_time=5) as receiver:
                 messages = []
                 for message in receiver:
                     messages.append(message)
@@ -243,7 +243,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
 
             session_id = str(uuid.uuid4())
             messages = []
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, 
+            with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               session_id=session_id, 
                                               receive_mode=ReceiveMode.ReceiveAndDelete, 
                                               max_wait_time=5) as session:
@@ -270,7 +270,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                     sender.send_messages(message)
 
             count = 0
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, 
+            with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               session_id=session_id, 
                                               max_wait_time=5) as session:
                 for message in session:
@@ -281,7 +281,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
 
             assert count == 10
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, 
+            with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               session_id=session_id, 
                                               max_wait_time=5) as session:
                 deferred = session.receive_deferred_messages(deferred_messages)
@@ -311,7 +311,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                 sender.send_messages(messages)
 
             count = 0
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, 
+            with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               session_id=session_id, 
                                               max_wait_time=5) as session:
                 for message in session:
@@ -322,7 +322,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
 
                 assert count == 10
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, 
+            with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               session_id=session_id, 
                                               max_wait_time=5) as session:
                 deferred = session.receive_deferred_messages(deferred_messages)
@@ -362,7 +362,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                     sender.send_messages(message)
 
             count = 0
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
                 for message in session:
                     deferred_messages.append(message.sequence_number)
                     print_message(_logger, message)
@@ -370,7 +370,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                     message.defer()
 
             assert count == 10
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, 
+            with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               session_id=session_id, 
                                               max_wait_time=5,
                                               receive_mode=ReceiveMode.ReceiveAndDelete) as session:
@@ -400,7 +400,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                     message = Message("Deferred message no. {}".format(i), session_id=session_id)
                     sender.send_messages(message)
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
                 count = 0
                 for message in session:
                     deferred_messages.append(message.sequence_number)
@@ -426,7 +426,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             servicebus_namespace_connection_string, logging_enable=False) as sb_client:
 
             session_id = str(uuid.uuid4())
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, 
+            with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               session_id=session_id, 
                                               max_wait_time=5, 
                                               prefetch_count=10) as receiver:
@@ -483,7 +483,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                 with sb_client.get_queue_receiver(servicebus_queue.name):
                     messages = sb_client.peek_messages(5)
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                 messages = receiver.peek_messages(5)
                 assert len(messages) == 5
                 assert all(isinstance(m, PeekedMessage) for m in messages)
@@ -492,7 +492,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                     with pytest.raises(AttributeError):
                         message.complete()
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id_2) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id_2) as receiver:
                 messages = receiver.peek_messages(5)
                 assert len(messages) == 3
 
@@ -508,7 +508,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             servicebus_namespace_connection_string, logging_enable=False) as sb_client:
 
             session_id = str(uuid.uuid4())
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, max_wait_time=5, session_id=session_id) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, max_wait_time=5, session_id=session_id) as receiver:
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     for i in range(5):
                         message = Message("Test message no. {}".format(i), session_id=session_id)
@@ -535,7 +535,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             session_id = str(uuid.uuid4())
             messages = []
             locks = 3
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, prefetch_count=10) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, prefetch_count=10) as receiver:
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     for i in range(locks):
                         message = Message("Test message no. {}".format(i), session_id=session_id)
@@ -592,7 +592,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
 
             renewer = AutoLockRenewer()
             messages = []
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ReceiveMode.PeekLock, prefetch_count=10) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ReceiveMode.PeekLock, prefetch_count=10) as receiver:
                 renewer.register(receiver.session, timeout=60, on_lock_renew_failure = lock_lost_callback)
                 print("Registered lock renew thread", receiver.session._locked_until_utc, utc_now())
                 with pytest.raises(SessionLockExpired):
@@ -630,7 +630,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             renewer._renew_period = 1
             session = None
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ReceiveMode.PeekLock, prefetch_count=10) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ReceiveMode.PeekLock, prefetch_count=10) as receiver:
                 session = receiver.session
                 renewer.register(session, timeout=5, on_lock_renew_failure=lock_lost_callback)
             sleep_until_expired(receiver.session)
@@ -657,7 +657,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                 message.session_id = session_id
                 sender.send_messages(message)
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                 messages = receiver.receive_messages(max_wait_time=10)
                 assert len(messages) == 1
 
@@ -681,7 +681,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                 message.session_id = session_id
                 sender.send_messages(message)
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                 messages = receiver.receive_messages(max_wait_time=10)
                 assert len(messages) == 1
                 print_message(_logger, messages[0])
@@ -697,7 +697,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                 with pytest.raises(SessionLockExpired):
                     receiver.session.renew_lock()
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                 messages = receiver.receive_messages(max_wait_time=30)
                 assert len(messages) == 1
                 print_message(_logger, messages[0])
@@ -717,7 +717,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
 
             session_id = str(uuid.uuid4())
             enqueue_time = (utc_now() + timedelta(minutes=2)).replace(microsecond=0)
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     content = str(uuid.uuid4())
                     message_id = uuid.uuid4()
@@ -754,7 +754,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             session_id = str(uuid.uuid4())
             enqueue_time = (utc_now() + timedelta(minutes=2)).replace(microsecond=0)
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, prefetch_count=20) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, prefetch_count=20) as receiver:
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     content = str(uuid.uuid4())
                     message_id_a = uuid.uuid4()
@@ -802,7 +802,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                 assert len(tokens) == 2
                 sender.cancel_scheduled_messages(tokens)
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                 messages = []
                 count = 0
                 while not messages and count < 13:
@@ -828,7 +828,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                     message = Message("Handler message no. {}".format(i), session_id=session_id)
                     sender.send_messages(message)
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5) as session:
                 assert session.session.get_state(timeout=5) == None
                 session.session.set_state("first_state", timeout=5)
                 count = 0
@@ -861,10 +861,10 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                         message = Message("Test message no. {}".format(i), session_id=session_id)
                         sender.send_messages(message)
             for session_id in sessions:
-                with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id) as receiver:
+                with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                     receiver.set_state("SESSION {}".format(session_id))
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE, max_wait_time=5, receive_mode=ReceiveMode.PeekLock) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE_SESSION, max_wait_time=5, receive_mode=ReceiveMode.PeekLock) as receiver:
                 current_sessions = receiver.list_sessions(updated_since=start_time)
                 assert len(current_sessions) == 5
                 assert current_sessions == sessions
@@ -892,7 +892,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                         message = Message("Test message no. {}".format(i), session_id=session)
                         sender.send_messages(message)
             for session in sessions:
-                with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session) as receiver:
+                with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session) as receiver:
                     receiver.set_state("SESSION {}".format(session))
 
                     current_sessions = receiver.list_sessions(updated_since=start_time)
@@ -913,7 +913,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
         def message_processing(sb_client):
             while True:
                 try:
-                    with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE, max_wait_time=5) as session:
+                    with sb_client.get_queue_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE_SESSION, max_wait_time=5) as session:
                         for message in session:
                             print("Message: {}".format(message))
                             messages.append(message)
@@ -959,7 +959,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                     message = Message("Handler message no. {}".format(i), session_id=session_id)
                     sender.send_messages(message)
 
-            with sb_client.get_queue_session_receiver(servicebus_queue.name, session_id=session_id, prefetch_count=0, max_wait_time=5) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, prefetch_count=0, max_wait_time=5) as receiver:
                 message = receiver.next()
                 assert message.sequence_number == 1
                 message.abandon()
@@ -984,7 +984,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                 message = Message(b"Sample topic message", session_id='test_session')
                 sender.send_messages(message)
 
-            with sb_client.get_subscription_session_receiver(
+            with sb_client.get_subscription_receiver(
                 topic_name=servicebus_topic.name,
                 subscription_name=servicebus_subscription.name,
                 session_id='test_session',